move normal aggregation to aggration file
